% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnv}
\alias{cnv}
\title{Sample Copy Number Alternation data}
\format{
A data frame with 709 genes and 109 variables:
\describe{
  \item{geneSymbol}{Gene name}
  \item{C3L.00081}{CNV dosage measurement for sample C3L.00081}
  \item{C3L.00415}{CNV dosage measurement for sample C3L.00415}
  ...
}
}
\usage{
cnv
}
\description{
A CNV dataset
cnv
}
\keyword{datasets}
