% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cov}
\alias{cov}
\title{Sample covariate data}
\format{
A data frame with 2 rows (covariates: age & female) and 108 variables:
\describe{
  \item{C3L.00081}{covariate measurements for sample C3L.00081}
  ...
  }
}
\usage{
cov
}
\description{
cov
}
\keyword{datasets}
