% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProFun.eFDR.R
\name{estFDR}
\alias{estFDR}
\title{Fast estimation of FDR without permutation procedure for one data type}
\usage{
estFDR(ProbPattern1y1x, grids = seq(0.01, 0.99, by = 0.01))
}
\arguments{
\item{ProbPattern1y1x}{Posterior probabilities for associations between one predictor data type and one outcome data type.}

\item{grids}{grids specify the searching grids to find significant associations}
}
\value{
\item{estFDR:}{Estimated FDR values by different posterior probability cutoffs considered in grids}
}
\description{
Fast estimation of FDR without permutation procedure for one data type
}
