% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitFDist.R
\name{fitFDist}
\alias{fitFDist}
\title{Internal function for calculating empirical bayes estimates}
\usage{
fitFDist(x, df1, covariate = NULL)
}
\arguments{
\item{x}{Input scale (e.g. variance of regression errors)}

\item{df1}{Degrees of freedom for input regression.}

\item{covariate}{Covariates}
}
\value{
\item{scale:}{Scale for rescaled t distribution.}
\item{df2:}{Degrees of freedom for rescaled t distribution.}
}
\description{
fitFDist is the function to obtain the empirical bayes estimate of n0 and s0^2 described on the top of the file. This function is extracted from limma package.
}
