% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProFun.detection.R
\name{iProFun.eFDR.1y}
\alias{iProFun.eFDR.1y}
\title{iProFun eFDR assessment for one outcome data type.}
\usage{
iProFun.eFDR.1y(
  reg.all,
  which.y,
  yList,
  xList,
  covariates,
  pi1,
  NoProbXIndex = NULL,
  permutate_number = 10,
  var.ID = c("Gene_ID"),
  var.ID.additional = NULL,
  seed = NULL
)
}
\arguments{
\item{reg.all}{The regression summary (unformatted) such as from iProFun.reg.}

\item{which.y}{The index for one outcome data type that the eFDR assessment is for.}

\item{yList}{yList is a list of data matrix for outcomes.}

\item{xList}{xList is a list of data matrix for predictors.}

\item{covariates}{covariates is a list of data matrix for covariate.}

\item{pi1}{pi1 is pre-specified prior of proportion of non-null statistics. It cane be a number in (0, 1) or a vector
of numbers with length of ylist.}

\item{NoProbXIndex}{NoProbXIndex allows users to provide the index for the predictor data type(s) that are not considered
for calculating posterior probabilities of association patterns.}

\item{permutate_number}{Number of permutation, default 10}

\item{var.ID}{var.ID gives the variable name (e.g. gene/protein name) to match different data types.
If IDs are not specified, the first columns will be considered as ID variable.}

\item{var.ID.additional}{var.ID.additional allows to output additional variable names from the input.
Often helpful if multiple rows (e.g. probes) are considered per gene to allow clear index of the rows.}

\item{seed}{seed allows users to externally assign seed to replicate results.}
}
\value{
\item{eFDR.grid:}{eFDR by the grid of posterior probability cutoffs.}
\item{fdr_cutPob:}{the cutoff values for pre-specified eFDR rate and the posterior probabilities for a pair of
 data types based on permutation.}
\item{No.Identified.filter:}{the number of identified variables for each  pair of data types.}
\item{No.Identified.no.filter:}{the number of identified variables for each  pair of data types.}
\item{Gene_fdr:}{A table summarizing the posterior probabilities (PostProb), the eFDR (eFDR.no.filter),
the significance under different criteria (nominal FDR, PostProb cutoffs and filter)  for each variable (e.g. gene)
under consideration.}
}
\description{
iProFun empirical false discovery rate (eFDR) assessment for one outcome data type.
}
