% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProFun.detection.R
\name{iProFun.eFDR}
\alias{iProFun.eFDR}
\title{iProFun eFDR assessment based on permutation for multiple outcome data type.}
\usage{
iProFun.eFDR(
  reg.all,
  yList,
  xList,
  covariates,
  pi1,
  NoProbXIndex = NULL,
  permutate_number = 10,
  var.ID = c("Gene_ID"),
  var.ID.additional = NULL,
  seed = NULL
)
}
\arguments{
\item{reg.all}{The regression summary (unformatted) such as from iProFun.reg.}

\item{yList}{yList is a list of data matrix for outcomes.}

\item{xList}{xList is a list of data matrix for predictors.}

\item{covariates}{covariates is a list of data matrix for covariate.}

\item{pi1}{pi1 is pre-specified prior of proportion of non-null statistics. It cane be a number in (0, 1) or a vector
of numbers with length of ylist.}

\item{NoProbXIndex}{NoProbXIndex allows users to provide the index for the predictor data type(s) that are not considered
for calculating posterior probabilities of association patterns.}

\item{permutate_number}{Number of permutation, default 10}

\item{var.ID}{var.ID gives the variable name (e.g. gene/protein name) to match different data types.
If IDs are not specified, the first columns will be considered as ID variable.}

\item{var.ID.additional}{var.ID.additional allows to output additional variable names from the input.
Often helpful if multiple rows (e.g. probes) are considered per gene to allow clear index of the rows.}

\item{seed}{seed allows users to externally assign seed to replicate results.}
}
\value{
\item{xName:}{Name of the predictors.}
\item{PostProb:}{The association probability for each gene on each data type.}
\item{Gene_efdr:}{The eFDR for each gene on each data type.}
}
\description{
iProFun empirical false discovery rate (eFDR) assessment based on permutation for multiple outcome data types.
}
