% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProFun.reg.1y.R
\name{iProFun.FWER}
\alias{iProFun.FWER}
\title{Identification with FWER}
\usage{
iProFun.FWER(Reg.Sum, FWER.Index = 0, filter = NULL, cutoff = 0.05)
}
\arguments{
\item{Reg.Sum}{Linear regression analysis summaries formatted in multi.omic.reg.summary or iProFun.reg.formatted.}

\item{FWER.Index}{Index the predictor data types that calculate FWER directly}

\item{filter}{filter is a vector with the same length of FWER.Index, taking values of 1, -1, 0 or NULL.
"NULL" is default imposes no filtering. 1" indicates that an association is considered for
significance only if its significant associations are positive across all outcome platforms. "-1" indicates
that an association is considered
for significance only if its significant associations are negative  across all outcome platforms. "0" indicates
that an association is considered for significance only if its significant association across all outcome
platforms preserve consistent directions (either positive or negative).}

\item{cutoff}{The cutoff value for declaring significance. The cutoff=0.05 in Default.}
}
\value{
It returns original p-values and FWER with and without filter
\item{pvalue:}{The p-value of linear regression with Student's t distribution}
\item{FWER.no.filter:}{FWER with Bonferroni correction of p-values}
\item{FWER.filter:}{Bonferroni correction of p-values among variables that pass biological filtering across data types.}
}
\description{
This function calculates FWER across all the variables in the same data types, directly
from regression results without considering posterior probabilities of association patterns.
This strategy is preferred for data types (e.g. somatic mutations) that have few variables that cannot reliably infer association patterns.
}
