% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProFun.reg.1y.R
\name{iProFun.prob}
\alias{iProFun.prob}
\title{Calculating posterior probabilities of association patterns}
\usage{
iProFun.prob(Reg.Sum, NoProbXIndex = NULL, pi1 = 0.05)
}
\arguments{
\item{Reg.Sum}{Linear regression analysis summaries formatted in multi.omic.reg.summary or iProFun.reg.formatted.}

\item{NoProbXIndex}{NoProbXIndex allows users to provide the index for the predictor data type(s) that are not considered
for calculating posterior probabilities of association patterns. NoProbXIndex = NULL (default): all data types in
predictors will be considered. Any 0 < NoProbXIndex <= length of xList:
indicates that the posterior probabilities of association patterns for the corresponding data type(s) are not calculated.}

\item{pi1}{pi1 is the pre-specified prior proportion of non-null statistics. It cane be a number in (0, 1) or a vector
of numbers with length of ylist.}
}
\value{
A list with the same length as xlist. Nested within each list, it contains
\item{NoComputation:}{No of variables considered for this predictor with all outcomes}
\item{Config:}{Corresponding association patterns. Total number 2^J for J outcome data types}
\item{Config.miss:}{Corresponding association patterns for variables (e.g. genes) that are missing in some outcome data type(s)}
\item{PostProb:}{Posterior probability for each predictor on each association pattern}
\item{PostProb.miss:}{Posterior probability for each predictor on each available association pattern  for variables (e.g. genes) that are missing in some outcome data type(s)}
\item{xName.miss:}{Predictor name for variables (e.g. genes) missing in some data type(s)}
\item{colocProb:}{Averaged posterior probability for predictor on each association pattern (missing data excluded)}
\item{Tstat_L:}{T statistics for each predictor on each outcome}
\item{D0:}{Estimated density under the null for predictor on each outcome}
\item{D1:}{Estimated density under the alternative for predictor on each outcome}
}
\description{
This is the core function of iProFun, and it calculates the posterior probabilities of association patterns from
regression summaries from multiple data types. It allows some data types from regression to opt out from calculating
posterior probabilities, which is suitable for the data types with very few genes (e.g. somatic mutation) that
borrowing information across variables is not reliable.
}
