% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProFun.reg.1y.R
\name{iProFun.reg}
\alias{iProFun.reg}
\title{Linear regression on all outcome data types (unformatted)}
\usage{
iProFun.reg(
  yList,
  xList,
  covariates,
  permutation.col = 0,
  var.ID = c("Gene_ID"),
  var.ID.additional = NULL,
  seed = NULL
)
}
\arguments{
\item{yList}{yList is a list of data matrix for outcomes.}

\item{xList}{xList is a list of data matrix for predictors.}

\item{covariates}{covariates is a list of data matrix for covariate.}

\item{permutation.col}{permutation.col provides the index of the data types that should be permuated.
permutation.col = 0 (default): no permuatation and analysis is on original data. 0 < permutate <= length of yList:
permuate the label of the corresponding data type in yList. For example, permutate =2, permute the y label of second
data matrix.}

\item{var.ID}{var.ID gives the variable name (e.g. gene/protein name) to match different data types.
If IDs are not specified, the first columns will be considered as ID variable.}

\item{var.ID.additional}{var.ID.additional  allows to output additional variables from the input.
Often helpful if multiple rows (e.g. probes) are considered per gene to allow clear index of results.}

\item{seed}{seed allows users to externally assign seed to replicate results.}
}
\value{
list with the same length as xlist. Nested within each list, it contains
\item{reg.out.list:}{reg.out.list returns the regression summary for each outcome data types as a list. Within each list,
see output of iProFun.reg.1y for details.}
}
\description{
Linear regression on all outcome data types with all types of DNA alterations (results not formatted for iProFun input)
}
