% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProFun.reg.R
\name{iProFun.reg.table}
\alias{iProFun.reg.table}
\title{Reformat the regression output into a table format. Each row is for a predictor-outcome pair.}
\usage{
iProFun.reg.table(reg.all, xType = NULL, yType = NULL)
}
\arguments{
\item{reg.all}{Output from iProFun.reg.}

\item{xType}{A vector of string for the data types of xList, such as "mutation", "CNV" and "methylation".}

\item{yType}{A vector of string for the data types of xList, such as "RNA", "protein" and "phospho".}
}
\value{
A table that includes gene ID, other gene info if provided, predictor data type, outcome data type,
estimate, se and p-value from Student's t-test, in a long format.
}
\description{
Reformat the regression output into a table format. Each row is for a predictor-outcome pair.
}
