% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProFun.prob.R
\name{iProFun.sum.prob.1y}
\alias{iProFun.sum.prob.1y}
\title{Summarize the association probabilities for an outcome.}
\usage{
iProFun.sum.prob.1y(prob, which.y = 1, NoProbXIndex = NULL)
}
\arguments{
\item{prob}{The posterior probabilieis of association patterns across multiple data types,
including association patterns with missing data in some of the data types. The output of iProFun.prob can be used here.}

\item{which.y}{The index for one outcome data type that the summation is for.}

\item{NoProbXIndex}{The index for the predictor data type(s) that are not considered
for calculating posterior probabilities of association patterns.}
}
\value{
\item{prob.all:}{The summarized posterior probabilities for all outcome variables in the data type of interest
(regardless of missingness in other data types).}
}
\description{
Summarize the association probabilities for an outcome across different association patterns.
The summation includes variables with missingness in some data types.
}
