% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{methy}
\alias{methy}
\title{Sample Methylation data}
\format{
A data frame with 1103 rows and 528 variables:
\describe{
  \item{Gene_ID}{Gene name}
  \item{TCGA-04-1331}{Methylation measurement for TCGA-04-1331}
  \item{TCGA-04-1332}{Methylation measurement for TCGA-04-1332}
  ...
}
}
\usage{
methy
}
\description{
A Methylation dataset
methy
}
\keyword{datasets}
