% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{Mstep}
\alias{Mstep}
\title{Internal function for EM algorithm M step}
\usage{
Mstep(oldB, n_trio, n_pattern)
}
\arguments{
\item{oldB}{Posterior probabilities for each trio}

\item{n_trio}{No. of trios}

\item{n_pattern}{No. of association patterns (=2^J).}
}
\description{
Internal function for EM algorithm M step
}
