% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProFun.reg.1y.R
\name{multi.omic.reg.summary}
\alias{multi.omic.reg.summary}
\title{Reformat regression summaries}
\usage{
multi.omic.reg.summary(reg.out.list, var.ID = var.ID)
}
\arguments{
\item{reg.out.list}{reg.out.list is the regression summaries from iProFun.reg.}

\item{var.ID}{var.ID gives the variable name (e.g. gene/protein name) to match different data types.}
}
\value{
list with the same length as xlist. Nested within each list, it contains
\item{betas_J:}{Coefficient estimate for predictors across J outcome data types}
\item{betas_se_J:}{Coefficent SE for predictors across J outcome data types}
\item{sigma2_J:}{Regrssion error term for predictors across J outcome data types}
\item{dfs_J:}{Regression degrees of freedom for predictors across J outcome data types}
\item{v_g_J:}{ (X^T X)^{-1} projection on predictors across J outcome data types}
\item{xName_J:}{Predictor name corresponds to each predictor-outcome pair across J outcome data types}
\item{yName_J:}{Outcome name corresponds to each predictor-outcome pair across J outcome data types}
}
\description{
Reformat regression summaries from iProFun.reg as input of iProFun analysis for association probabilities and discoveries
}
