% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mut}
\alias{mut}
\title{Sample Mutation data}
\format{
A data frame with 13 genes and 109 variables:
\describe{
  \item{geneSymbol}{Gene name}
  \item{C3L.00081}{Mutation (Yes=1; No=0) for sample C3L.00081}
  \item{C3L.00415}{Mutation (Yes=1; No=0) sample C3L.00415}
  ...
}
}
\usage{
mut
}
\description{
A Mutation dataset
mut
}
\keyword{datasets}
