% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phospho}
\alias{phospho}
\title{Sample Phospho data}
\format{
A data frame with 4168 phosphosites and 110 variables:
\describe{
  \item{geneSymbol}{Gene name}
  \item{id}{Phospho site identification}
  \item{C3L.02665}{Phospho measurement for C3L.02665}
  ...
}
}
\usage{
phospho
}
\description{
A phospho dataset
phospho
}
\keyword{datasets}
