% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protein}
\alias{protein}
\title{Sample Protein data}
\format{
A data frame with 551 genes and 109 variables:
\describe{
  \item{geneSymbol}{Gene name}
  \item{C3L.02665}{Protein measurement for sample C3L.02665}
  \item{C3L.01663}{Protein measurement for sample C3L.01663}
  ...
}
}
\usage{
protein
}
\description{
A protein dataset
methy
}
\keyword{datasets}
