% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rna}
\alias{rna}
\title{Sample mRNA data}
\format{
A data frame with 699 genes and 109 variables:
\describe{
  \item{geneSymbol}{Gene name}
  \item{C3L.00081}{mRNA measurement for sample C3L.00081}
  \item{C3L.00415}{mRNA measurement for sample C3L.00415}
  ...
}
}
\usage{
rna
}
\description{
A mRNA dataset
rna
}
\keyword{datasets}
