% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rna_pc_1_3}
\alias{rna_pc_1_3}
\title{First 3 principle components for rna data}
\format{A data frame with 68 rows and 4 variables:
\describe{
  \item{subject_id}{Subject ID name}
  \item{rna_pc1}{First principle component for rna data}
  \item{rna_pc2}{Second principle component for rna data}
  \item{rna_pc3}{Third principle component for rna data}
}}
\usage{
rna_pc_1_3
}
\description{
First 3 principle components for rna data
rna_pc_1_3
}
\keyword{datasets}
