% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapOmicsQTL_function.R
\name{Estep}
\alias{Estep}
\title{consider matrixStats::rowMins() or rpgm::rowMins()}
\usage{
Estep(cl, oldpi, n_trio, n_pattern, q, density_0, density_1)
}
\description{
consider matrixStats::rowMins() or rpgm::rowMins()
}
