% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProFun.eFDR.R
\name{iProFun.eFDR}
\alias{iProFun.eFDR}
\title{iProFun eFDR assessment based on permutation for multiple outcome data type.}
\usage{
iProFun.eFDR(
  reg.all,
  yList,
  xList,
  covariates,
  pi1,
  NoProbXIndex = NULL,
  filter = NULL,
  permutate_number = 10,
  grids = seq(0.01, 0.99, by = 0.01),
  fdr = 0.1,
  PostCut = 0.75,
  var.ID = c("Gene_ID"),
  var.ID.additional = c("phospho_ID", "Hybridization", "chr"),
  seed = NULL
)
}
\arguments{
\item{reg.all}{The regression summary (unformatted) such as from iProFun.reg.}

\item{yList}{yList is a list of data matrix for outcomes.}

\item{xList}{xList is a list of data matrix for predictors.}

\item{covariates}{covariates is a list of data matrix for covariate.}

\item{pi1}{pi1 is pre-specified prior of proportion of non-null statistics. It cane be a number in (0, 1) or a vector
of numbers with length of ylist.}

\item{NoProbXIndex}{NoProbXIndex allows users to provide the index for the predictor data type(s) that are not considered
for calculating posterior probabilities of association patterns.}

\item{filter}{filter is a vector with the same length of xList, taking values of 1, -1, 0 or NULL.
"NULL" is default imposes no filtering. 1" indicates that an association is considered for
significance only if its significant associations are positive across all outcome platforms. "-1" indicates
that an association is considered
for significance only if its significant associations are negative  across all outcome platforms. "0" indicates
that an association is considered for significance only if its significant association across all outcome
platforms preserve consistent directions (either positive or negative).}

\item{permutate_number}{Number of permutation, default 10}

\item{grids}{Grids specify the searching grids for significant associations}

\item{fdr}{norminal false discover rate, default at 0.1}

\item{PostCut}{PostCut specifies minimal posterior probabilty cutoff to be considered as
significant, default as 0.75}

\item{var.ID}{var.ID gives the variable name (e.g. gene/protein name) to match different data types.
If IDs are not specified, the first columns will be considered as ID variable.}

\item{var.ID.additional}{var.ID.additional allows to output additional variable names from the input.
Often helpful if multiple rows (e.g. probes) are considered per gene to allow clear index of the rows.}

\item{seed}{seed allows users to externally assign seed to replicate results.}
}
\value{
\item{eFDR.grid:}{eFDR by the grid of posterior probability cutoffs.}
\item{fdr_cutPob:}{the cutoff values for pre-specified eFDR rate and the posterior probabilities for a pair of
 data types based on permutation.}
\item{No.Identified.filter:}{the number of identified variables for each  pair of data types.}
\item{No.Identified.no.filter:}{the number of identified variables for each  pair of data types.}
\item{Gene_fdr:}{A table summarizing the posterior probabilities (PostProb), the eFDR (eFDR.no.filter),
the significance under different criteria (nominal FDR, PostProb cutoffs and filter)  for each variable (e.g. gene)
under consideration.}
}
\description{
iProFun empirical false discovery rate (eFDR) assessment based on permutation for multiple outcome data type.
}
