% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProFun.detection.R
\name{iProFun.FWER}
\alias{iProFun.FWER}
\title{Calculate FWER}
\usage{
iProFun.FWER(reg.all, FWER.Index = 0)
}
\arguments{
\item{reg.all}{Linear regression analysis summaries from iProFun.reg.}

\item{FWER.Index}{Index the predictor data types that calculate FWER directly}
}
\value{
It returns original p-values and FWER with and without filter
\item{pvalue:}{The p-value of linear regression with Student's t distribution}
\item{FWER:}{The FWER with Bonferroni correction for all genes}
\item{xName:}{The gene name of the predictors}
}
\description{
This function calculates FWER across all the outcome variables for the same predictor data type
(e.g. mutation), directly from regression results without considering posterior probabilities of association patterns.
This strategy is preferred for data types that have few genes that cannot reliably infer association patterns.
}
