% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProFun.R
\name{iProFun}
\alias{iProFun}
\title{iProFun Integrative analysis}
\usage{
iProFun(ylist = list(rna, protein, phospho), xlist = list(cna, methy),
  covariates = list(rna_pc_1_3, protein_pc_1_3, phospho_pc_1_3),
  pi = rep(0.05, 3), permutate = 0)
}
\arguments{
\item{ylist}{mRNA, global protein and phosphoprotein}

\item{xlist}{CNA and methylation}

\item{covariates}{sets of covariates adjusted in the regression analyses}

\item{pi}{prior probability}

\item{permutate}{whether to permuate or not. 1 = permuatte the first column(mRNA), 2 = permutate the second column, 3 = permutate the third column}
}
\value{
list with 3 components
\item{Marginal Probability:}{marginal posterior probability for each pattern}
\item{Gene Posterior Probability:}{posterior probability for each pattern per methylation site}
\item{Beta:}{Beta coefficients for each gene-level multiple linear regression}
}
\description{
iProFun Integrative analysis
}
\examples{
iprofun_result <- iprofun()
}
