% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProFun.R
\name{iProFun}
\alias{iProFun}
\title{iProFun Integrative analysis}
\usage{
iProFun(ylist = list(rna, protein, phospho), xlist = list(cna, methy),
  covariates = list(rna_pc_1_3, protein_pc_1_3, phospho_pc_1_3),
  pi = rep(0.05, 3), permutate = 0)
}
\arguments{
\item{ylist}{ylist is a list of data matrix. Each matrix is a data type or platform that
one would like to analyze as an outcome. The matrix is formated as each outcome variable
(such as gene) as a row, and each sample as a column. Multiple data matrixs need to have
shared sample ID and variable name for integrative iProFun analysis. Example of ylist is
a list of mRNA, global protein and phosphoprotein.}

\item{xlist}{xlist is a list of data matrix. Each matrix is a data type or platform that
one would like to analyze as an predictor. The matrix is formated as each outcome variable
(such as gene) as a row, and each sample as a column. Multiple data matrixs need to have
shared sample ID and variable name for integrative iProFun analysis. Example of xlist is
a list of cna and methylation.}

\item{covariates}{covariates is a list of data matrix. Each matrix is a data type or platform that
one would like to adjust as a covariate. The matrix is formated as each outcome variable
as a row, and each sample as a column. Multiple data matrixs need to have
shared sample ID and variable name for integrative iProFun analysis. Example of covariates is
a list of principle components.}

\item{pi}{prior probability}

\item{permutate}{whether to permuate or not. 1 = permuatte the first column(mRNA), 2 = permutate the second column, 3 = permutate the third column}
}
\value{
list with 3 components
\item{Marginal Probability:}{marginal posterior probability for each pattern}
\item{Gene Posterior Probability:}{posterior probability for each pattern per methylation site}
\item{Beta:}{Beta coefficients for each gene-level multiple linear regression}
}
\description{
iProFun Integrative analysis
}
\examples{
iprofun_result <- iprofun(ylist = list(rna, protein, phospho), xlist = list(cna, methy), covariates = list(rna_pc_1_3, protein_pc_1_3, phospho_pc_1_3), pi = rep(0.05, 3))
}
