% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProFun.R
\name{iProFun}
\alias{iProFun}
\title{iProFun Integrative analysis}
\usage{
iProFun(ylist, xlist, covariates, pi = rep(0.05, 3), permutate = 0,
  ID = NULL, x.ID = NULL, y.ID = NULL, sub.ID.common = "TCGA",
  colum.to.keep = c("phospho_ID", "Hybridization", "chr"),
  missing.rate.filter = NULL, verbose = T)
}
\arguments{
\item{ylist}{ylist is a list of data matrix. Each matrix is a data type or platform that
one would like to analyze as an outcome. The matrix is formated as each outcome variable
(such as gene) as a row, and each sample as a column. Multiple data matrixs need to have
shared sample ID and variable ID for integrative iProFun analysis. If more than one rows
are availble for one variable ID (e.g. multiple phosphosites per gene with gene ID as
variable ID), only the row that are most promising is selected via ANOVA for the final
output. Example of ylist is a list of mRNA, global protein and phosphoprotein.}

\item{xlist}{xlist is a list of data matrix. Each matrix is a data type or platform that
one would like to analyze as an predictor. All data types or platforms are simutaneouosly
considered in the same regression. The matrix is formated as each predictor variable
(such as gene) as a row, and each sample as a column. Multiple data matrixs need to have
shared sample ID and variable ID for integrative iProFun analysis. If more than one rows
are availble for one variable ID (e.g. multiple methylation sites per gene with gene ID as
variable ID), all rows are considered in the regression to account for the effects of each
other. Example of xlist is a list of CNA and DNA methylation.}

\item{covariates}{covariates is a list of data matrix. This list should have the same length
as ylist. For the regression on the ith outcome, the ith covariates matrix contains the variables
that to be adjusted in the regression. The matrix is formated as each covariate variable
as a row, and each sample as a column. Multiple data matrixs need to have
shared sample ID and variable ID for integrative iProFun analysis. Example of covariates is
a list of principle components.}

\item{pi}{pi is pre-specified priori of proportion of non-null statistics in each set of regression.
iProFun is insensitive to the mis-specification of the priori within a reasonable range.}

\item{permutate}{whether to permuate certain data type/platform or not. permutate = 0 (default):
no permuatation, analysis on original data. 0 < permutate <= length of ylist: permuate the label of
the corresponding data type in ylist. For example, permutate =2, permute the y label of second data matrix (e.g.protein)}

\item{ID}{ID gives the variable name (e.g. gene/protein name) to match different data types and consider integrative
analysis on. If IDs are not specified, the first columns will be considered as ID variable. iProFun only
applies with a fair number of overlapping variables are overlappe across all data types.}

\item{x.ID}{x.ID specifies variable name for xlist. It provides the flexible option to use
different names for xlist and ylist.}

\item{y.ID}{y.ID specifies variable name for ylist. It provides the flexible option to use
different names for xlist and ylist.}

\item{sub.ID.common}{sub.ID.common specifies the common string for samples (e.g. TCGA). If not
specified, all variables except IDs and colum.to.keep will be searched for overlapping samples.}

\item{colum.to.keep}{colum.to.keep allows to output additional variables except IDs in the
final results. Often helpful if multiple rows (e.g. probes) are considered per ID variable
(e.g. gene) to allow clear specification of each result.}

\item{missing.rate.filter}{missing.rate.filter allows filtering out variables with high missing
rate in data.}

\item{verbose}{verbose=T print out the progress of iProFun analysis.}
}
\value{
list with the same length as xlist. Nested within each list, it contains
\item{betas_J:}{Coefficient estimate for predictor on each outcome}
\item{betas_se_J:}{Coefficent SE for predictor on each outcome}
\item{sigma2_J:}{Regrssion error term for predictor on each outcome}
\item{dfs_J:}{Regression degrees of freedom for predictor on each outcome}
\item{v_g_J:}{ (X^T X)^{-1} projection on predictor on each outcome}
\item{xName_J:}{Predictor name corresponds to each predictor outcome pair}
\item{yName_J:}{Outcome name corresponds to each predictor outcome pair}
\item{NoComputation:}{No of variables considered for this predictor with all outcomes}
\item{Config:}{Corresponding association patterns. Total number 2^J for J outcome data types}
\item{PostProb:}{Posterior probability for each predictor on each association pattern}
\item{colocProb:}{Averaged posterior probability for predictor on each association pattern}
\item{Tstat_L:}{T statistics for each predictor on each outcome}
\item{D0:}{Estimated density under the null for predictor on each outcome}
\item{D1:}{Estimated density under the alternative for predictor on each outcome}
}
\description{
iProFun Integrative analysis
}
\examples{
iprofun_result <- iProFun(ylist = list(rna, protein, phospho),
xlist = list(cna, methy), covariates = list(rna_pc_1_3, protein_pc_1_3, phospho_pc_1_3), pi = rep(0.05, 3))
}
