% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProFun_permutate.R
\name{iProFun_permutate}
\alias{iProFun_permutate}
\title{iProFun false discovery rate assessment based on permutations}
\usage{
iProFun_permutate(ylist = list(rna, protein, phospho),
  xlist = list(cna, methy), covariates = list(rna_pc_1_3,
  protein_pc_1_3, phospho_pc_1_3), pi = rep(0.05, 3),
  permutate_number = 10, fdr = 0.1, posterior = 0.75)
}
\arguments{
\item{ylist}{mRNA, global protein and phosphoprotein}

\item{xlist}{CNA and methylation}

\item{covariates}{sets of covariates adjusted in the regression analyses}

\item{pi}{prior probability}

\item{permutate_number}{Number of permutation, default 10}

\item{fdr}{False Discover Rate, default as 0.1}

\item{posterior}{Minimal Posterior Probabilty Cutoff, default as 0.75}
}
\value{
list with 3 components
\item{Posterior Probability Cutoff:}{the cutoff values for each group based on permutation}
\item{iProFun Result:}{A table indicating whether a gene is identified by iProFun or not}
\item{iProFun Result (Negative/Positive):}{A table indicating whether a gene is identified by iProFun or not and whether the association is positive or not}
}
\description{
iProFun false discovery rate assessment based on permutations
}
\examples{
iprofun_permutate_result <- iProFun_permutate(permutate_number = 1)
}
