% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProFun_permutate.R
\name{iprofun_permutate}
\alias{iprofun_permutate}
\title{Calculate the empirical FDR using permutation}
\usage{
iprofun_permutate(ylist = list(rna, protein, phospho),
  xlist = list(cna, methy), covariates = list(rna_pc_1_3,
  protein_pc_1_3, phospho_pc_1_3), pi = rep(0.05, 3),
  permutate_number = 10)
}
\arguments{
\item{ylist}{mRNA, global protein and phosphoprotein}

\item{xlist}{CNA and methylation}

\item{covariates}{sets of covariates adjusted in the regression analyses}

\item{pi}{prior probability}

\item{permutate_number}{Number of permutation, default 10}
}
\description{
Calculate the empirical FDR using permutation
}
\examples{
iprofun_permutate
}
