% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProFun_permutate.R
\name{iProFun_permutate}
\alias{iProFun_permutate}
\title{iProFun false discovery rate assessment based on permutations}
\usage{
iProFun_permutate(ylist, xlist, covariates, pi, permutate_number = 10,
  fdr = 0.1, PostCut = 0.75, grids = seq(0.01, 0.99, by = 0.01),
  filter = NULL, seed = .Random.seed[1], ID = NULL, x.ID = NULL,
  y.ID = NULL, sub.ID.common = "TCGA",
  colum.to.keep = c("phospho_ID", "Hybridization", "chr"),
  missing.rate.filter = NULL)
}
\arguments{
\item{ylist}{ylist is a list of data matrix. Each matrix is a data type or platform that
one would like to analyze as an outcome. The matrix is formated as each outcome variable
(such as gene) as a row, and each sample as a column. Multiple data matrixs need to have
shared sample ID and variable name for integrative iProFun analysis. Example of ylist is
a list of mRNA, global protein and phosphoprotein.}

\item{xlist}{xlist is a list of data matrix. Each matrix is a data type or platform that
one would like to analyze as an predictor. The matrix is formated as each outcome variable
(such as gene) as a row, and each sample as a column. Multiple data matrixs need to have
shared sample ID and variable name for integrative iProFun analysis. Example of xlist is
a list of cna and methylation.}

\item{covariates}{covariates is a list of data matrix. Each matrix is a data type or platform that
one would like to adjust as a covariate. The matrix is formated as each outcome variable
as a row, and each sample as a column. Multiple data matrixs need to have
shared sample ID and variable name for integrative iProFun analysis. Example of covariates is
a list of principle components.}

\item{pi}{prior probability}

\item{permutate_number}{Number of permutation, default 10}

\item{fdr}{false discover rate, default as 0.1}

\item{PostCut}{PostCut specifies minimal posterior probabilty cutoff to be considered as
significant, default as 0.75}

\item{grids}{grids specify the searching grids to find significant associations}

\item{filter}{filter is a vector with the same length of xlist, taking values of 1, -1, 0 or NULL.
"NULL" is default imposes no filtering. 1" indicates that an association is considered for
significance only if in positive directions across all outcome platforms. "-1" indicates
that an association is considered
for significance only if in negative directions across all outcome platforms. "0" indicates
that an association is considered for significance only if association across all outcome
platforms preserve consistent directions (either positive or negative).}

\item{seed}{seed allows users to externally assign seed to replicate results.}
}
\value{
list with 4 components
\item{fdr_grid:}{FDR by posterior probability cutoffs considered in grids}
\item{fdr_cutPob:}{the cutoff values for prespecified FDR rate and posterior probability for each X and Y pairs based on permutation}
\item{NoIdentified:}{the number of identified variables for each X and Y pair}
\item{Gene_fdr:}{A table indicating whether a gene is significantly identified by iProFun or not. "1" indicates the gene
is significantly identified; "0" indicates the gene is not.}
\item{PostProb_original:}{The posterior probabilities in iProFun analysis with input data.}
\item{PostProb_perm:}{The posterior probabilities with permutated data. for the i-the permutation for j-th Y and k-th X,
the posterior probabilities is saved inPostProb_perm[[k]][[j]][[i]].}
}
\description{
iProFun false discovery rate assessment based on permutations
}
\examples{
iprofun_permutate_result <- iProFun_permutate(ylist = list(rna, protein, phospho),
xlist = list(cna, methy), covariates = list(rna_pc_1_3, protein_pc_1_3, phospho_pc_1_3),
pi = rep(0.05, 3), permutate_number = 1, fdr = 0.1, PostCut = 0.75, filter <- c(1,0),
grids = c(seq(0.75, 0.99, 0.01), seq(0.991, 0.999, 0.001), seq(0.9991, 0.9999, 0.0001)),
seed=123)
}
