% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protein}
\alias{protein}
\title{Sample Protein data}
\format{
A data frame with 1103 rows and 528 variables:
\describe{
  \item{Gene_ID}{Gene name}
  \item{TCGA-04-1331}{Protein measurement for TCGA-04-1331}
  \item{TCGA-04-1332}{Protein measurement for TCGA-04-1332}
  ...
}
}
\usage{
protein
}
\description{
A protein dataset
methy
}
\keyword{datasets}
