% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rna}
\alias{rna}
\title{Sample mRNA data}
\format{
A data frame with 676 rows and 526 variables:
\describe{
  \item{Gene_ID}{Gene name}
  \item{TCGA-04-1331}{mRNA measurement for TCGA-04-1331}
  \item{TCGA-04-1332}{mRNA measurement for TCGA-04-1332}
  ...
}
}
\usage{
rna
}
\description{
A mRNA dataset
rna
}
\keyword{datasets}
